added authRoutes.js in backend - all routing for logging vs. signing in
