+++ backend/authRoutes.js
@@ -1,0 +1,89 @@
+// authRoutes.js
+const express = require('express');
+const bcrypt = require('bcrypt');
+const { PrismaClient } = require('@prisma/client');
+const { OAuth2Client } = require('google-auth-library');
+
+const prisma = new PrismaClient();
+const client = new OAuth2Client(process.env.GOOGLE_CLIENT_ID);
+
+const router = express.Router();
+
+// Register a new user
+router.post('/register', async (req, res) => {
+  const { email, username, password } = req.body;
+
+  // Hash the password before storing it
+  const hashedPassword = await bcrypt.hash(password, 10);
+
+  try {
+    const newUser = await prisma.user.create({
+      data: {
+        email,
+        username,
+        password: hashedPassword,
+      },
+    });
+    res.json({ success: true, user: newUser });
+  } catch (error) {
+    console.error('Error registering user:', error);
+    res.status(500).json({ success: false, error: 'Error registering user' });
+  }
+});
+
+// Login an existing user
+router.post('/login', async (req, res) => {
+  const { email, password } = req.body;
+
+  try {
+    const user = await prisma.user.findUnique({ where: { email } });
+
+    if (!user) {
+      return res.status(404).json({ success: false, error: 'User not found' });
+    }
+
+    const passwordMatch = await bcrypt.compare(password, user.password);
+
+    if (!passwordMatch) {
+      return res.status(401).json({ success: false, error: 'Invalid credentials' });
+    }
+
+    res.json({ success: true, user });
+  } catch (error) {
+    console.error('Error logging in:', error);
+    res.status(500).json({ success: false, error: 'Error logging in' });
+  }
+});
+
+// Login user with Google
+router.post('/google-login', async (req, res) => {
+  const { token } = req.body;
+
+  try {
+    const ticket = await client.verifyIdToken({
+      idToken: token,
+      audience: process.env.GOOGLE_CLIENT_ID,
+    });
+    const payload = ticket.getPayload();
+    const { email, name, picture } = payload;
+
+    let user = await prisma.user.findUnique({ where: { email } });
+
+    if (!user) {
+      user = await prisma.user.create({
+        data: {
+          email,
+          username: name,
+          picture,
+        },
+      });
+    }
+
+    res.json({ success: true, user });
+  } catch (error) {
+    console.error('Error with Google login:', error);
+    res.status(500).json({ success: false, error: 'Error with Google login' });
+  }
+});
+
+module.exports = router;
